requirements.txt
anyascii==0.3.2
asgiref==3.6.0
bcp47==0.0.4
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dashboard==0.0.6
Django==4.1.8
django-allauth==0.54.0
django-appconf==1.0.5
django-enumchoicefield==3.0.1
django-filter==22.1
django-mirage-field==1.4.0
django-modelcluster==6.0
django-nullmailer==0.3
django-permissionedforms==0.1
django-ranged-response==0.2.0
django-recaptcha
django-taggit==3.1.0
django-treebeard==4.7
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
load_dotenv
html5lib==1.1
idna==3.4
l18n==2021.3
oauthlib==3.2.2
openpyxl==3.1.2
Pillow==9.5.0
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.17.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
telepath==0.3
tqdm==4.65.0
Unidecode==1.3.6
urllib3==1.26.15
wagtail==5.1.1
wagtail-django-recaptcha
wagtail-generic-chooser==0.5.1
wagtailgmaps
wagtailmedia
wagtailvideos
wagtailmenus==3.1.8
wagtail_pdf_view
webencodings==0.5.1
uwsgi
unicode
Willow==1.6.2
